refactor: repo structure and drawing function
