+++ src/tncdr/evolutors/stabilizer/pauli_string.py
@@ -1,5 +1,4 @@
 from copy import copy
-from typing import List
 
 # CONVERSION TABLES
