+++ src/tncdr/evolutors/stabilizer/random_clifford.py
@@ -1,8 +1,17 @@
 import random as rnd
 from typing import Generator
 
-from tncdr.stabilizer_mps.pauli_string import Pauli
-from  tncdr.stabilizer_mps.tableaus import Tableau, CNOT, SWAP, H, S, X, Z
+from tncdr.evolutors.stabilizer.pauli_string import Pauli
+from tncdr.evolutors.stabilizer.tableaus import (
+    Tableau,
+    CNOT,
+    SWAP,
+    S,
+    H,
+    X,
+    Z,
+)
+
 
 def random_pauli(n:int)->Pauli:
     """
