+++ src/tncdr/targets/ansatze.py
@@ -1,4 +1,3 @@
-from abc import ABC, abstractmethod
 from dataclasses import dataclass
 
 import numpy as np
