+++ tests/test_entropies_calculation.py
