+++ tests/test_pauli.py
